simVis: RFProp thickness is now always in # height steps (SIM-10297, SIM-7246)
 Release Note: N/A

 Notes: thickness in units of height steps is more clear, matches the data better.

 Test Environment: Windows 10, VC14, Debug, 64-bit

 Test: areps and builder data

 Review: 19924

+++ SDK/simVis/RFProp/Profile.h
@@ -77,11 +77,11 @@ public:
   /** Sets the DataProvider for this Profile */
   void setDataProvider(CompositeProfileProvider* dataProvider);
 
-  /** Gets the display thickness, in meters for this Profile. */
-  float getDisplayThickness() const;
+  /** Gets the display thickness, in # of height steps for this Profile. */
+  unsigned int getDisplayThickness() const;
 
-  /** Sets the display thickness, in meters for this Profile. This setting effects the DRAWMODE_3D DisplayMode, as well as 3D Points and 3D Texture. */
-  void setDisplayThickness(float displayThickness);
+  /** Sets the display thickness, in # of height steps for this Profile. This setting effects the DRAWMODE_3D DisplayMode, as well as 3D Points and 3D Texture. */
+  void setDisplayThickness(unsigned int displayThickness);
 
   /** Gets the alpha of this Profile */
   float getAlpha() const;
@@ -221,8 +221,8 @@ protected:
   /** Bearing of the profile in radians */
   double bearing_;
 
-  /** Profile display thickness */
-  float displayThickness_;
+  /** Profile display thickness, in # height steps */
+  unsigned int displayThickness_;
   /** Height of vertical slots */
   double height_;
   /** Half of the beam width in radians */
